Keep the extension view container object
Before this commit, the extension view container object was being
destroyed and created whenever the user moved from and then back to the
Extension Views.  This caused the equivalent of a refresh.  Now, the
object is created once and its display is changed whenever the user
moves from and to the Extension Views.